+++ grpc_router/core/models.py
@@ -1,4 +1,20 @@
 from dataclasses import dataclass
+import datetime
+from enum import Enum
+from typing import Optional
+
+
+class HealthCheckType(Enum):
+    NONE = 0
+    ACTIVE_CLIENT = 1
+    PASSIVE_CLIENT = 2
+
+
+class HealthStatus(Enum):
+    UNKNOWN = 0
+    GOOD = 1
+    WARNING = 2
+    ERROR = 3
 
 
 @dataclass
@@ -12,10 +28,28 @@ class ConfigOptions:
 
 
 @dataclass
+class ServiceHealthStatus:
+    status: HealthStatus
+    description: str
+    timestamp: datetime.datetime
+
+
+@dataclass
 class Service:
     service_id: str
     host: str
     port: int
     region: str
     slots: int
+    health_check_type: HealthCheckType
     service_token: str
+
+    health_status: Optional[ServiceHealthStatus] = None
+
+
+    def set_health_status(self, status: HealthStatus, description: str, timestamp: datetime.datetime) -> None:
+        self.health_status = ServiceHealthStatus(
+            status=status,
+            description=description,
+            timestamp=timestamp
+        )